feat: app and db permissions in settings tab
