+++ src/ui/shared/environment/index.ts
@@ -1,2 +1,3 @@
 export * from "./environment-list";
 export * from "./backup-retention-policy-view";
+export * from "./perms";
